+++ includes/tokenization.h
@@ -6,7 +6,7 @@
 /*   By: marsoare <marsoare@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/29 16:41:18 by marsoare          #+#    #+#             */
-/*   Updated: 2024/10/27 20:15:57 by marsoare         ###   ########.fr       */
+/*   Updated: 2024/10/28 13:02:13 by marsoare         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,27 +42,42 @@ typedef struct s_token
 
 //check_token_type.c
 t_token_type	token_type(char *value);
+
 //create_token_lst.c
 void			tokenize_input(t_shell *shell, char *input);
+
 //expand_tokens.c
 void			expand_tokens(t_shell *shell);
+
 //t_list	*tokenize_input(const char *input);
 int				handle_word_token(t_shell *shell, char *input, int i);
 int				handle_redir(t_shell *shell, char *input, int i);
 int				handle_pipe(t_shell *shell, char *input, int i);
 int				handle_quotes(t_shell *shell, char *input, int i);
+
 //ft_joinstrs.c
 int				join_strs(t_shell *shell, char **str, char *input, int i);
-int				handle_quoted_str(t_shell *shell, char **str, char *input, int i);
-int				handle_nonquoted_str(t_shell *shell, char **str, char *input, int i);
+int				h_quoted_str(t_shell *shell, char **str, char *input, int i);
+int				h_nonquoted_str(t_shell *shell, char **str, char *input, int i);
+
 //handle_expand.c
 int				handle_expand(t_shell *shell, char *input, int i);
+int				prcs_expansion(t_shell *shell, char **str, char *input, int i);
+int				expand_unquoted(t_shell *shell, char **str, char *input, int i);
+int				expand_quoted(t_shell *shell, char **str, char *input, int i);
+
+//handle_expand_utils.c
+char			*ft_strjoin_char(char *str, char c);
+
 //lexer.c
 void			lexer(t_shell *shell, char	*input);
+
 //remove_qutes.c
-void	remove_quotes(t_shell *shell);
+void			remove_quotes(t_shell *shell);
+
 //set_token_pos.c
 void			set_token_pos(t_list *lst);
+
 //tokenize_utils.c
 bool			in_quotes(char *input, int i);
 bool			ft_ismeta(char *str, int i);
